[WIP] Changed LevelBuilder to a function in LevelManager. Level building now clears any previous physics colliders/shapes.
