+++ src/server/levels/manager.js
@@ -1,9 +1,75 @@
 const fs = require("fs").promises;
 const log = require("../../log");
 const levelIO = require("../../level/level-io");
-const levelBuilder = require("./level-builder");
+const physics = require("../../physics/manager");
 
 const levelManager = function() {
+	// Level builder helper function
+	let _transformFromEntity = function(entity) {
+		let transform = new physics.ammo.btTransform();
+		transform.setIdentity();
+		transform.setOrigin(
+			new physics.ammo.btVector3(
+				entity.position.x,
+				entity.position.y,
+				entity.position.z
+			)
+		);
+		transform.setRotation(
+			new physics.ammo.btQuaternion(
+				entity.rotation.x,
+				entity.rotation.y,
+				entity.rotation.z,
+				entity.rotation.w
+			)
+		);
+
+		return transform;
+	};
+
+	// Parses level data and loads it as their appropriate physics colliders
+	let _buildLevel = function(levelData) {
+		// Clear previous level data, if any
+		physics.world.clearColliders();
+		physics.clearShapes();
+
+		for(let key in levelData.models) {
+			let model = levelData.models[key];
+			if(model.convexData) {
+				physics.createConvexShape(key, model.convexData);
+			}
+			if(model.concaveData) {
+				physics.createConcaveShape(key, model.concaveData.vertices, model.concaveData.indices);
+			}
+		}
+
+		physics.world.setGravity(levelData.gameplay.gravity);
+
+		for (let worldObjectUuid in levelData.worldObjects) {
+			let worldEntity = levelData.worldObjects[worldObjectUuid];
+			let prefab = levelData.prefabs[levelData.worldObjects[worldObjectUuid].prefab];
+
+			for (let prefabEntityUuid in prefab.entities) {
+				let prefabEntity = prefab.entities[prefabEntityUuid];
+
+				if(prefabEntity.type !== "collider") continue;
+
+				let worldEntityTransform = _transformFromEntity(worldEntity);
+				let prefabEntityTransform = _transformFromEntity(prefabEntity);
+
+				// Clone the transform because op_mul modifies the transform it is called on
+				let transform = new physics.ammo.btTransform();
+				transform.setIdentity();
+				transform.setOrigin(worldEntityTransform.getOrigin());
+				transform.setRotation(worldEntityTransform.getRotation());
+
+				transform.op_mul(prefabEntityTransform); // Modifies "transform"
+
+				physics.world.createCollider(prefabEntity, transform);
+			}
+		}
+	};
+
 	return {
 		availableLevels: null,
 		currentLevel: null,
@@ -107,7 +173,7 @@ const levelManager = function() {
 					if(!level) {
 						return Promise.reject(`Level "${levelName}" is invalid.`);
 					}
-					levelBuilder(level);
+					_buildLevel(level);
 					this.currentLevel = level;
 					return level;
 				})
